+++ src/components/sidebar/sidebar_bottom.svelte
@@ -1,0 +1,5 @@
+<script>
+    import UserMenu from "~/components/userMenu.svelte";
+</script>
+
+<UserMenu />